pyproject.toml
[tool.poetry]
name = "data-loader"
version = "0.1.0"
description = ""
authors = ["Pedro Fuenmayor <p.fuenmayor@ampeersenergy.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.5"}
pydantic = "^2.9.2"
strawberry-graphql = {extras = ["fastapi"], version = "^0.248.1"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
